Move host FFI and unified RPC context from bindings/ into src/api
CLI, tests, and every language binding already share c4_rpc_ctx_t; keep that layer with the C core as libapi / libapi_ffi instead of compiling the wrappers in each consumer.
