+++ bindings/emscripten/CMakeLists.txt
@@ -72,8 +72,8 @@ configure_file(
   @ONLY
 )
 
-add_executable(c4w ems.c ../colibri_common.c)
-target_link_libraries(c4w prover verifier eth_verifier)
+add_executable(c4w ems.c)
+target_link_libraries(c4w api prover verifier eth_verifier)
 set_target_properties(c4w PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY ${EMSCRIPTEN_OUTPUT_DIR}
     LINK_FLAGS "${EMC_PROPS}"
