+++ src/CMakeLists.txt
@@ -28,6 +28,9 @@ if(PROVER)
     add_subdirectory(prover)
 endif()
 add_subdirectory(chains)
+if(VERIFIER AND PROVER)
+    add_subdirectory(api)
+endif()
 if(CLI)
     add_subdirectory(cli)
 endif()
